dev bitmap for mm

+++ include/common/types.h
@@ -14,8 +14,8 @@ typedef double float64_t;
 
 #define NULL 0
 
-#define true 1
-#define false 0
-typedef int bool;
+#define true (uint8_t)1
+#define false (uint8_t)0
+typedef uint8_t bool;
 
 #endif
